# src/controllers/types/product-request.type.ts
export interface ProductCreateRequest {
  name: string;
  price: number;
  category: string;
  stock: number;
}

// Add this
export interface ProductUpdateRequest {
  name?: string;
  price?: number;
  category?: string;
  stock?: number;
}